+++ app/src/at-hungtran/java/asiantech/internship/summer/HungActivity.java
@@ -1,0 +1,37 @@
+package asiantech.internship.summer;
+
+import android.content.Intent;
+import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+
+public class HungActivity extends AppCompatActivity {
+    private Button btn;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        initView();
+        initListener();
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+    }
+    private void initView(){
+        btn = findViewById(R.id.idBtn);
+    }
+    private void initListener() {
+        btn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(HungActivity.this, LayoutActivity.class);
+                startActivity(intent);
+            }
+        });
+    }
+}
