+++ app/src/at-hungtran/res/drawable/border1.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <!-- Set màu nền-->
+    <solid
+        android:color="#1659df" >
+    </solid>
+
+    <!-- và đây là bán kính đường tròn ở 4 góc -->
+    <corners
+        android:radius="50dp" >
+    </corners>
+
+</shape>
