+++ app/src/at-hungtran/res/layout/activity_main.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout
+    xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
+    android:layout_height="match_parent">
+    <Button
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/idBtn"
+        android:text="@string/viewgroup"
+        android:layout_centerHorizontal="true"/>
+</RelativeLayout>
