+++ app/src/at-hungtran/res/values/strings.xml
@@ -8,6 +8,6 @@
     <string name="loadding">loadding</string>
     <string name="todo">TODO</string>
     <string name="conent">conent</string>
-    <string name="viewgroup">ViewGroup</string>
+    <string name="viewgroup">W1-D3 View and ViewGroup</string>
     <string name="profile">Profile</string>
 </resources>
