+++ app/src/main/java/asiantech/internship/summer/MainActivity.java
@@ -1,39 +1,12 @@
 package asiantech.internship.summer;
 
-import android.content.Intent;
-import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
 
 public class MainActivity extends AppCompatActivity {
-    private Button btn;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        initView();
-        initListener();
-
-
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-    }
-    private void initView(){
-        btn = findViewById(R.id.idBtn);
-    }
-    private void initListener() {
-        btn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(MainActivity.this, LayoutActivity.class);
-                startActivity(intent);
-            }
-        });
     }
 }
